JetBrains: Don't handle alt+enter in Java if it happens inside the web view (https://github.com/sourcegraph/sourcegraph/pull/36374)
@vdavid noticed here that the key listener restructuring caused the alt+enter keybind to always be handled in the Java code. This is because we have added the same listener code for both events from the Java side and events from the web view side.

This PR fixes this by adding a boolean argument to differentiate the two. We already have working handlers for alt+enter on the web view and it is conceptually important that the web view can handle the data that should be opened in that case because of possible data races.