JetBrains: Cody: implement multiline autocomplete for the editor (https://github.com/sourcegraph/sourcegraph/pull/54848)
Fixes https://github.com/sourcegraph/sourcegraph/pull/54095
Fixes https://github.com/sourcegraph/sourcegraph/pull/54140 (there still is stuff to be done, but I do not think it
should be tracked under a single, huge umbrella ticket, more details
below)

The first thing is, there is no longer any need to cut a suggestion if
it includes the current caret line suffix, as we can now render both
before and after it.
It's also integrated nicely upon applying the autocomplete.


https://github.com/sourcegraph/sourcegraph/assets/18601388/a182df52-d49d-43b5-b8da-00eb009b4034

The second thing is, the rendering is now handled for multiline
suggestions.
The following short recording is coming from Cody, not a mockup.


https://github.com/sourcegraph/sourcegraph/assets/18601388/f2d748dd-37f0-4469-acf1-7c2b65fce485

Now, the prompts we have don't really have all the complex stuff for
getting multiline stuff, but it seems we are still getting multiline
autocomplete suggestions for some contexts. Those will nowe be rendered
and be usable in the plugin.
As getting our `prompt_library` up-to date with the VS Code's multiline
fireworks would be an enormous effort, I do believe we should wait for
the agent integration instead. That's also why I don't think that part
should be tracked under https://github.com/sourcegraph/sourcegraph/pull/54140 - a separate ticket would be better, I
think.

Now, for the stuff that doesn't work:
- if the suggestion includes not just the current caret line suffix, but
also subsequent lines in the editor, we won't merge them and paste a
duplicate after the current line instead
- I did not manage to replicate this with the current Cody autocomplete
suggestions, so I'm not even sure if it's possible
  - it should be ok to fix it in a subsequent PR, just to be safe
- if we get a multiline suggestion which does not include the current
line suffix, it's not handled properly and the resulting autocomplete
might be completely invalid
- we should render over the current line suffix in that case, which is
not yet done (and I did not yet manage to find a way to hack around it)
- again, I did not manage to replicate a case like that with the current
Cody autocomplete suggestions, so I'm not even sure if it's likely to
get autocomplete like that
- it's something we should probably fix as soon as possible, but I don't
think it's a large enough reason to hide multiline behind a feature flag
- if we do decide a feature flag is necessary, this would be the main
reason

NOTE: **I could hide the multiline stuff behind a feature flag if it is
deemed necessary, but I do think it should be fine to leave it the way
it is. Hiding the feature might rob us of valuable early user feedback,
so it's something to consider. Try playing with it and et me know if
it's necessary.**

## Test plan
- green CI 🟢 
- please play with the feature a bit and try breaking it before
accepting the PR